+++ test/spec/idx/authenticate.ts
@@ -1967,7 +1967,7 @@ describe('idx/authenticate', () => {
 
         });
 
-        it('can get Okta Verify link via SMS', async () => {
+        it('can get Okta Verify link via SMS (proceed via channel)', async () => {
           const {
             authClient,
             selectAuthenticatorResponse,
@@ -2026,6 +2026,73 @@ describe('idx/authenticate', () => {
           });
 
           expect(enrollmentChannelDataSmsResponse.proceed).toHaveBeenCalled();
+        });
+
+        it('can get Okta Verify link via SMS (proceed via authenticator)', async () => {
+          const {
+            authClient,
+            selectAuthenticatorResponse,
+            enrollPollResponse,
+            enrollmentChannelDataSmsResponse,
+            successResponse
+          } = testContext;
+
+          chainResponses([
+            selectAuthenticatorResponse,
+            enrollPollResponse,
+            enrollmentChannelDataSmsResponse,
+            successResponse
+          ]);
+
+          jest.spyOn(mocked.introspect, 'introspect')
+            .mockResolvedValueOnce(selectAuthenticatorResponse)
+            .mockResolvedValueOnce(enrollPollResponse)
+            .mockResolvedValueOnce(enrollPollResponse) // submit enrollment channel
+            .mockResolvedValueOnce(enrollmentChannelDataSmsResponse);
+
+          jest.spyOn(enrollPollResponse, 'proceed');
+          jest.spyOn(enrollmentChannelDataSmsResponse, 'proceed');
+
+          await authenticate(authClient, {
+            authenticator: AuthenticatorKey.OKTA_VERIFY
+          });
+          let res = await proceed(authClient, {
+            step: 'select-enrollment-channel'
+          });
+          const { options } = res.nextStep!;
+          expect(options).toContainEqual({
+            label: 'SMS',
+            value: 'sms'
+          });
+
+          res = await proceed(authClient, {
+            authenticator: {
+              id: 'string',
+              channel: 'phoneNumber'
+            }
+          });
+          const { inputs } = res.nextStep!;
+
+          expect(enrollPollResponse.proceed).toHaveBeenCalledWith(
+            'select-enrollment-channel',
+            expect.objectContaining({ authenticator: expect.objectContaining({ channel: 'phoneNumber' }) })
+          );
+
+          expect(inputs).toContainEqual({
+            name: 'phoneNumber',
+            label: 'Phone Number',
+            required: true,
+            type: 'string',
+          });
+
+          await proceed(authClient, {
+            phoneNumber: '+1234'
+          });
+
+          expect(enrollmentChannelDataSmsResponse.proceed).toHaveBeenCalledWith(
+            'enrollment-channel-data',
+            expect.objectContaining({ phoneNumber: '+1234' })
+          );
         });
       });
     });